Initial cut of gha-exporter
This initial cut produces a binary that works for a single repo. It
pulls the workflow runs and jobs from GitHub Actions via its API v3
(http) and sums the step/job/workflow run times for successful steps,
jobs and workflows and accumulates them into prometheus metrics as
described in `metrics.md`.

It loops and regularly collects run times, keeping track of what it has
seen so as to not query unnecessarily.

There's still a bunch of TODOs, needs some tests and could do with some
build/ci scripts to produce an OCI image.